requirements.txt
graphene>=2.1.3
graphene-django>=2.1.0
pycountry>=18.5
Statistics>=1.0.3
